refactor: rearrange file structure based on the official angular style guide
