+++ src/app/app.component.html
@@ -1,2 +1,2 @@
-<sg-app-sidebar></sg-app-sidebar>
+<sidebar-sidebar></sidebar-sidebar>
 <router-outlet></router-outlet>
