+++ src/app/app.module.ts
@@ -1,3 +1,4 @@
+import { DashboardModule } from './dashboard/dashboard.module';
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
@@ -5,66 +6,25 @@ import { AppRoutingModule } from './app-routing.module';
 
 import { AppComponent } from './app.component';
 
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { MatInputModule } from '@angular/material/input';
-import { MatCardModule } from '@angular/material/card';
-import { MatButtonModule } from '@angular/material/button';
-import { MatIconModule } from '@angular/material/icon';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatNativeDateModule, MAT_DATE_LOCALE } from '@angular/material/core';
-import { MatDatepickerModule } from '@angular/material/datepicker';
-import { MatDialogModule } from '@angular/material/dialog';
-import { MatSelectModule } from '@angular/material/select';
-import { MatSlideToggleModule } from '@angular/material/slide-toggle';
-import { MatExpansionModule } from '@angular/material/expansion';
-import { DragDropModule } from '@angular/cdk/drag-drop';
-
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { LoginComponent } from './login/login.component';
-import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { SidebarComponent } from './sidebar/sidebar.component';
-import { StudentsComponent } from './students/students.component';
-import { NewStudentDialogComponent } from './new-student-dialog/new-student-dialog.component';
-import { DeleteStudentDialogComponent } from './delete-student-dialog/delete-student-dialog.component';
-import { OkDialogComponent } from './ok-dialog/ok-dialog.component';
-import { NewLessonDialogComponent } from './new-lesson-dialog/new-lesson-dialog.component';
+import { CoreModule } from './core/core.module';
+import { SharedModule } from './shared/shared.module';
+import { StudentsModule } from './students/students.module';
+import { LoginModule } from './login/login.module';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    DashboardComponent,
-    LoginComponent,
-    PageNotFoundComponent,
-    SidebarComponent,
-    StudentsComponent,
-    NewStudentDialogComponent,
-    DeleteStudentDialogComponent,
-    OkDialogComponent,
-    NewLessonDialogComponent,
-  ],
+  declarations: [AppComponent],
   imports: [
     BrowserModule,
-    AppRoutingModule,
     HttpClientModule,
     BrowserAnimationsModule,
-    ReactiveFormsModule,
-    FormsModule,
-    // Material Imports
-    MatInputModule,
-    MatCardModule,
-    MatButtonModule,
-    MatIconModule,
-    MatFormFieldModule,
-    MatDatepickerModule,
-    MatNativeDateModule,
-    MatDialogModule,
-    MatSelectModule,
-    MatSlideToggleModule,
-    MatExpansionModule,
-    DragDropModule,
+    CoreModule,
+    SharedModule,
+    DashboardModule,
+    StudentsModule,
+    LoginModule,
+    AppRoutingModule,
   ],
-  providers: [{ provide: MAT_DATE_LOCALE, useValue: 'hu-HU' }],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
