+++ src/app/dashboard/dashboard.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { DashboardRoutingModule } from './dashboard-routing.module';
+import { DashboardComponent } from './dashboard.component';
+import { WeeklyCalendarComponent } from './weekly-calendar/weekly-calendar.component';
+import { SharedModule } from '../shared/shared.module';
+
+
+@NgModule({
+  declarations: [DashboardComponent, WeeklyCalendarComponent],
+  imports: [
+    CommonModule,
+    DashboardRoutingModule,
+    SharedModule
+  ]
+})
+export class DashboardModule { }
