+++ src/app/shared/material/material.module.ts
@@ -1,0 +1,48 @@
+import { NgModule } from '@angular/core';
+
+import { MatInputModule } from '@angular/material/input';
+import { MatCardModule } from '@angular/material/card';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatNativeDateModule, MAT_DATE_LOCALE } from '@angular/material/core';
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { MatDialogModule } from '@angular/material/dialog';
+import { MatSelectModule } from '@angular/material/select';
+import { MatSlideToggleModule } from '@angular/material/slide-toggle';
+import { MatExpansionModule } from '@angular/material/expansion';
+import { DragDropModule } from '@angular/cdk/drag-drop';
+
+@NgModule({
+  declarations: [],
+  imports: [
+    MatInputModule,
+    MatCardModule,
+    MatButtonModule,
+    MatIconModule,
+    MatFormFieldModule,
+    MatNativeDateModule,
+    MatDatepickerModule,
+    MatDialogModule,
+    MatSelectModule,
+    MatSlideToggleModule,
+    MatExpansionModule,
+    DragDropModule,
+  ],
+  exports: [
+    MatInputModule,
+    MatCardModule,
+    MatButtonModule,
+    MatIconModule,
+    MatFormFieldModule,
+    MatNativeDateModule,
+    MatDatepickerModule,
+    MatDialogModule,
+    MatSelectModule,
+    MatSlideToggleModule,
+    MatExpansionModule,
+    DragDropModule,
+  ],
+  providers: [{ provide: MAT_DATE_LOCALE, useValue: 'hu-HU' }],
+})
+export class MaterialModule {}
