+++ src/app/students/students-routing.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { StudentsComponent } from './students.component';
+import { AuthGuard } from '../core/services/auth.guard';
+
+const routes: Routes = [
+  {
+    path: 'students',
+    component: StudentsComponent,
+    canActivate: [AuthGuard],
+  },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class StudentsRoutingModule {}
